Add import() capability to jjs commandline tool
Added a custom module loader to the commandline tool. The builtin module loader does not
work and is difficult to configure. The new loader is not to spec, but it works well enough for
file specifier uses cases (., .. and absolute). The builtin loader will need to be rewritten at
some point.

import() now works from stdin, regular js files and module files.

Patched broken utility js in 262 tests.