Partial Fix for 32-bit Builds
In some build configurations, the variables used to read the stack
pointer and checks stack limit exceeded get moved around. When
this happens, rangeerror exceptions can be thrown in tests.

This is a partial fix that works in some configurations, but does not
work with some 32 bit debug and usan builds.