Use Sensible Build Defaults
JJS has a lot of build options. With the current defaults, including JJS in a project (via cmake, amalgam, zig, etc)
requires fiddling with the build flags to get proper error messages and other stuff for embedding the engine into
an application.

For now, the CI build and test options will continue to use the old defaults. Some of the OOM tests depend on the
specific vm configuration. Will review at a later time.