(WIP) Rework Context
In the old code, the vm heap and jjs context were fixed at compile time. If you wanted to change the behavior,
you could create an external context from the port. I added the ability to set heap size at run time using the
external context feature. This worked for awhile, but there have been lots of bugs and testing issues related
to the external context. Ultimately, there were too many compile time switches to manage and get good
test coverage.

This change removes the concept of an external context. The context itself is now internal to JJS. The vm
heap can be configured to be a compile time setting or runtime configurable (not both). The default is runtime
configurable. The downstream heap and stack related code is now the same. Now, the vm heap, gc and stack
can be configured dynamically through jjs_init. And, you can even provide a memory block to be used for the
heap.

Everything related to the external context is gone. jjs_init has been reworked to return a status code and take
in settings. The test compile flags have looked arbitrary and it is a game of whack-a-mole to get the system
working on multiple platforms. I believe this was related to the code path permutations the external context
implementation caused. I found and fixed quite a few sanitizer bugs in the tests and a couple in the heap
code.