Turn off stack limit check by default
The feature takes a pointer to a local variables to calculate
stack limit exceeded. This does not work with asan:
https://github.com/google/sanitizers/wiki/AddressSanitizerUseAfterReturn#compatibility
In practice, on some compilers and configurations it works and others it does not. The
stack checking gets messed up on other configurations like 32 bit release (lto/non lto)
gcc builds without any sanitizers.

Now that the configuration flags are more sane and the test infrastructure is doing
better coverage. In the previous incarnation of the project, it looks like they were
playing with build configuration flags until the CI test suite worked. The gaps in
between were being ignored.

The stack limit is not a priority for JJS nor the projects that depend on JJS. By
default, stack limit check will be disabled. The code is still there, but the CI
will not test it. If you want the feature in your project, that is your responsibility
to test your projects configuration.