Add value ownership semantics
JS value references can be passed to an api with a new ownership enum. Keep means
the caller keeps ownership. Move means the caller transfers ownership to the api. On
success or error, the api will release a moved reference.

The old apis have not been changed. The semantic is for the new annex apis (for now).

This is not perfect, but the options for cleaning up resources in C are limited and I am
dealing with a legacy api. This scheme seems to make the code easier to write.