realpath
posix 2008 fixed realpath to take NULL for the resolved buffer and alloc the path memory
internally. though, some linux flavors still have other realpath bugs. at compile time,
try to select the best realpath available and fallback to the old buggy version.

in the future, remove the fallback and have the user choose a realpath impl for their
platform, but some support stuff is needed to make that easier.