Fix sanpshots when line info enabled
When line info is on, the line info is disabled in the binary, but
a 4 byte block is allocated for the line info pointer. The resulting
snapshot may or may not run on a vm, depending on several factors. When
line info is off, snapshot generation does not have the 4 byte block
and the snapshot may or may not run.

The byte code status flags have been exhausted, so I cannot patch this
with another flag. Instead, the 4 byte block is always allocated. The
compiled code has a flag to use it or not. The flag is for compiled code
saved to a snapshot. With some adjustments, the snapshots work whether
line info is on/off at snapshot generation or snapshot run.

The next step is to serialize line info and source name (?) into the
snapshot, but I am not sure if that is valuable (the original file
may be gone).

Note: Looking at this code, I am not sure there is value in line info
being a comp time flag. It seems to add complexity for a saving a
negligible amount of code and some minor if checks in parsing.