Create a custom executable for the test262 harness.
The main jjs commandline tool does too much and test262 specific stuff is not something
uses would need or use. And, the jjs cli interface is not what the test harness needs. The
harness was doing black magic to get imports to work properly. Now, the test harness
reads the source and pipes it into jjs-test262 with some test specific flags.

The new tool should be faster, but the python harness spends a lot of time reading the
test directory, excludes xml, reading each test and compiling a list of tests to run. That
looks like it can be improved in the future.