jjs:test
Add a dedicated jjs:test package to the cli. It exports a test method similar to node's test
runner. It supports sync tests, async tests and skipping tests. Tests must require or import
jjs:test manually. The package is available for the "test" cli command.

The cli reporter is minimal. The reporting is done by run-tests.py. In the future, the cli should
implement a better reporter.