Parser Performance
A 262 test, that ran 65000 evals in a tight loop, took 25 seconds + to run on
debug and 11 seconds on release. This change reduces that particular test time
to 200-300 ms and appears to greatly improve parser performance.

JJS was originally built to not call malloc/free for any operations. All memory
vm related or tempoary go through the vm heap. This makes sense for platforms
and use cases where dynamic allocations are not available or undesirable. However,
going through the vm heap for interstitial operations (module resolution, sorting,
concating strings, parsing state, etc) puts pressure on the gc. The slow test
was running into this issue because parsing, even for small strings, needs to
create alot of small objects to track state.

For module resolution, a scratch buffer was introduced to handle temporary
allocations for path manipulation. JJS can continue to be configured to run
without dynamic allocations and write to a single memory block. A section
of the block serves an arena-like allocator for temporary allocations and
the other section is the vm heap.

All parsing paths now use the scratch buffer for temporary state. Huge
performance improvement.

Note: The GC, heap alloc and heap free are super slow. Needs investiation.
Note: Many other places where vm heap is being used for temporary work.
      Those need to use the scratch buffer.