Performance Fixes
During parsing, JS literals (strings, numbers, bigint) are converted to JS values and
put into a cache for performance and bookkeeping purposes. The cache is very slow. It
created a literal JS value, checked for it's existence in the cache and if the literal
was already cached, the value would be destroyed (kinda defeats the purpose of the
cache). Lots of work and gc churn. The cache itself is a linked list and was traversed
at O(n). Depending on the source code, this cache and its constribution to gc churn
were the dominating performance benchmarks.

The creation of objects for lookups was remedied. There was also an insidious memory
leak that was found and fixed. JJS converts string literals to numbers automatically,
but the behavior is different depending on the number. Numbers in a certains range
were getting allocated by the cache, but never cleaned up. I think this was happening
at a low enough level where the memory leak detection never triggered.

The linked list was replaced with a custom hash set. The implementation needs some
work, but the performance results are great. VM memory pools are used to avoid
hitting the very slow heap.

In local tests, time to parse jslint.mjs was cut in half.