Performance
Improved performance of keyword compare in the lexer. For some reason it
was doing a binary search to check if an token was a keyword. gperf had
ok results, but the code it generates is difficult to deal with in the
current code base and it would be annoying to integrate into the build
process. I implemented a customer hash/matcher that outperforms gperf,
at least on my machine.

This is ultimately a small improvement to the parsing process, as most
of the time is in the parser.