Performance
Strings consisting of numbers are a special case. Internally, when the chars
of these strings are needed, they are generated on the fly into a buffer. In
many cases, this buffer was dynamically allocated. Since the max buffer
size is known and all the get char use cases are local to a function, use a
stack allocated buffer. The GC is not involved, but there was overhead with
allocating and deallocating these buffers adds up.