Performance
More work on literal storage. The backing hashset inserts and gets now have 90% less collisions
due to a better probe. The comparison seems not as efficient as possible, but for now, that is a
minor issue. The literal storage hashet actually uses less memory than the old linked list based
literal store. The linked list node had 2 pointers and each hashset node only has an ecma value.
With a 70% threshold on the hashset, that is a 20% memory savings for the literal pool itself.

Another issue is that the hashset does hit re-grows in the test environment. I set the initial lit
pool size at 1024, but it may need to be tuned and configurable for the user.