Performance Fix
The vm heap allocator is slow. For small allocations, there can be 4-8 bytes of overhead (more?). For many test scripts, 60%
or more of the allocations are 32 bytes or less. When lots of small allocations stick around, the vm heap lists get expensive to
traverse. Performance measurements become dominated by the heap working through all of this. There is a memory pool that
was outside of the heap, but it appeared to be adding to the problems.

The fix is to add a cell allocator to the vm heap for allocations <= 32 bytes. The environment does do small allocations of 8 or 16
bytes, but the current vm added overhead. 32 byte cells actually save quite a bit of memory. In the future, things can be optimized
to better utilize the memory. The minimum of 32 byte allocations makes everything much simpler. I was able to completely get
rid of the pool manager.

The cells are paged using 1024 as a default. This may need some tuning, as loading all pack modules requires 4 pages. The
pages are not freed until context shutdown. This also will need to be addressed, as this is not great for workflows that have
bursts of object creation.

For larger allocations, the slow path is still being used. As that accumulates more objects, it still gets slow. Still needs to be fixed.